Fixed a bug with pagination parameters being set to a constant and extending functionality

+++ Project/[Project]CheatSheet/[Project]CheatSheet/Features/Resources/Services/ResourceService.cs
@@ -16,7 +16,7 @@
 
     public class ResourceService : IResourceService
     {
-        private const int ResourcesPerPage = 12;
+        private const byte ResourcesPerPage = 12;
         private readonly CheatSheetDbContext context;
         private readonly ICurrentUser currentUserService;
         private readonly IMapper mapper;
